# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "消防巡检管理系统 API",
    "description": "基于 Node.js + Express + MongoDB 的消防设备巡检、报警处理、安全管理系统 API 文档。\n\n## 认证方式\n所有私有接口需在请求头中携带 Authorization: Bearer <token>。\n\n## 数据范围\n部分接口受数据范围权限控制：all（全部）/ department（本部门）/ self（仅本人）/ none（无权限）。",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "开发环境"
    },
    {
      "url": "/",
      "description": "当前服务器"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/inspections": {
      "get": {
        "tags": ["Inspections"],
        "summary": "获取巡检列表（分页）",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10,
              "maximum": 100
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["pending", "in_progress", "completed", "overdue", "cancelled"]
            }
          },
          {
            "name": "inspectionType",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["daily", "weekly", "monthly", "quarterly", "annual", "special"]
            }
          },
          {
            "name": "assignedTo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "post": {
        "tags": ["Inspections"],
        "summary": "创建巡检计划",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "inspectionType",
                  "planStartTime",
                  "planEndTime",
                  "checkItems"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "inspectionType": {
                    "type": "string",
                    "enum": ["daily", "weekly", "monthly", "quarterly", "annual", "special"]
                  },
                  "planStartTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "planEndTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "assignedTo": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "mongodb-id"
                    }
                  },
                  "devices": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "locations": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  },
                  "checkItems": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    },
                    "minItems": 1
                  },
                  "description": {
                    "type": "string"
                  },
                  "remark": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功"
          }
        }
      }
    },
    "/api/inspections/stats": {
      "get": {
        "tags": ["Inspections"],
        "summary": "获取巡检统计",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "返回按状态/类型/结果统计"
          }
        }
      }
    },
    "/api/inspections/{id}": {
      "get": {
        "tags": ["Inspections"],
        "summary": "获取巡检详情",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "put": {
        "tags": ["Inspections"],
        "summary": "更新巡检计划",
        "description": "仅 pending 状态可修改",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          },
          "400": {
            "description": "已开始的巡检不能修改"
          }
        }
      },
      "delete": {
        "tags": ["Inspections"],
        "summary": "删除巡检",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          },
          "400": {
            "description": "正在执行的巡检不能删除"
          }
        }
      }
    },
    "/api/inspections/{id}/start": {
      "put": {
        "tags": ["Inspections"],
        "summary": "开始执行巡检",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/api/inspections/{id}/complete": {
      "put": {
        "tags": ["Inspections"],
        "summary": "提交巡检结果",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["result"],
                "properties": {
                  "result": {
                    "type": "string",
                    "enum": ["normal", "abnormal", "partial"]
                  },
                  "findings": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  },
                  "location": {
                    "type": "string"
                  },
                  "remark": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/api/inspections/{id}/review": {
      "put": {
        "tags": ["Inspections"],
        "summary": "审核巡检结果",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["reviewResult"],
                "properties": {
                  "reviewResult": {
                    "type": "string",
                    "enum": ["approved", "rejected"]
                  },
                  "reviewComment": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/api/inspections/{id}/cancel": {
      "put": {
        "tags": ["Inspections"],
        "summary": "取消巡检",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
...
    "/api/reports/inspections": {
      "get": {
        "tags": ["Reports"],
        "summary": "获取巡检报表",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "返回按状态、结果、执行人统计"
          }
        }
      }
    },
...
  }
}